Introduced solution in main directory. Tmp dir for subtree is removed
