--- solution/.gitignore
@@ -1,3 +1,0 @@
-*.o
-*.exe
-test.cpp
